Get Vico compiling on the 10.8 SDK and update Nu to 2.1.

+++ app/SFTPConnection.m
@@ -712,7 +712,9 @@ resp2txt(int type)
 	const void *ptr;
 	NSUInteger len;
 
-	switch (event) {
+	// cast to int because ViStreamEventWriteEndEncountered is not declared
+	// part of the enum, and clang don't like that
+	switch ((int)event) {
 	case NSStreamEventNone:
 	case NSStreamEventOpenCompleted:
 	default:
